+++ frontend/src/scenes/ProfileScene/index.js
@@ -5,7 +5,7 @@ import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
 import {
   getPlayer, getPlayerGames, appendToPlayerGames, getPlayerDeltaStatistic, getPlayerGamesCountStatistic, getRelations,
-  getRelationsDashboard
+  getRelationsDashboard, getCurrent
 } from '../../actions';
 import {withRouter} from 'react-router-dom';
 import ProfileMainInfo from '../../components/profile-main-info';
@@ -19,9 +19,13 @@ import {
 
 class ProfileScene extends Component {
 
+  state = {
+    playerId: this.props.match.params.id
+  };
+
   componentDidMount() {
-    const {match, actions} = this.props;
-    const playerId = match.params.id;
+    const {actions} = this.props;
+    const {playerId} = this.state;
     actions.getPlayer(playerId);
     actions.getPlayerDeltaStatistic(playerId);
     actions.getPlayerGames(playerId);
@@ -31,8 +35,8 @@ class ProfileScene extends Component {
   }
 
   componentDidUpdate() {
-    const {player, actions, match} = this.props;
-    const playerId = match.params.id;
+    const {player, actions} = this.props;
+    const {playerId} = this.state;
 
     if (player && player.id !== +playerId) {
       actions.getPlayer(playerId);
@@ -52,32 +56,34 @@ class ProfileScene extends Component {
     if (!player) return null;
     if (player.id !== +playerId) return null;
 
-    const mappedGames = playerGames.list.map(game => (
-      {
-        ...game.gameDto,
-        delta: game.delta,
-        won: game.won,
-      }
-    ))
+    const mappedGames = playerGames.list.map(game => ({
+      ...game.gameDto,
+      delta: game.delta,
+      won: game.won,
+    }));
 
     const isCurrent = !!currentUser && (+playerId === currentUser.id);
     const isMobile = window.outerWidth <= MediaViews.MOBILE;
     return (
+
       <Content>
         <ProfileMainInfo countGames={player.countGames} rated={player.rated} rating={player.rating}
-                         id={this.props.match.params.id} goalsAgainst={player.goalsAgainst} goalsFor={player.goalsFor}
-                         bestPartner={relationsDashboard && relationsDashboard.bestPartner}
+                         id={this.state.playerId} goalsAgainst={player.goalsAgainst} goalsFor={player.goalsFor}
+                         bestPartner={relationsDashboard && relationsDashboard.bestPartner} iconPath={player.iconPath}
                          worstPartner={relationsDashboard && relationsDashboard.worstPartner}
                          favoritePartner={relationsDashboard && relationsDashboard.favoritePartner}
-												 fullName={player.fullName} isCurrent={isCurrent} countLosses={player.countLosses}
+                         fullName={player.fullName} isCurrent={isCurrent} countLosses={player.countLosses}
                          countWins={player.countWins} winningPercentage={player.winningPercentage}
                          currentLossStreak={player.currentLossesStreak} currentWinStreak={player.currentWinningStreak}
-                         longestLossStreak={player.longestLossesStreak} longestWinStreak={player.longestWinningStreak}/>
+                         longestLossStreak={player.longestLossesStreak} longestWinStreak={player.longestWinningStreak}
+                         getPlayer={actions.getPlayer} getCurrent={actions.getCurrent} email={player.email}/>
         <PlayerRelations relations={relations.list} isMobile={isMobile}/>
-        <ChartStatistics ratingStatistic={ratingStatistic} gamesCountStatistic={gamesCountStatistic} isMobile={isMobile}/>
+        <ChartStatistics ratingStatistic={ratingStatistic} gamesCountStatistic={gamesCountStatistic}
+                         isMobile={isMobile}/>
         <PlayerGames games={mappedGames ? mappedGames : []} appendToGames={actions.appendToPlayerGames}
                      totalCount={playerGames.totalCount} isMobile={isMobile} playerId={playerId}/>
       </Content>
+
     );
   }
 }
@@ -93,15 +99,21 @@ const mapStateToProps = (state) => { // eslint-disable-line no-unused-vars
     relationsDashboard: state.player.relationsDashboard,
   };
   return props;
-}
+};
 const mapDispatchToProps = (dispatch) => {
   const actions = {
-    getPlayer, getPlayerGames, appendToPlayerGames, getPlayerGamesCountStatistic, getPlayerDeltaStatistic, getRelations,
-    getRelationsDashboard
+    getPlayer,
+    getPlayerGames,
+    appendToPlayerGames,
+    getPlayerGamesCountStatistic,
+    getPlayerDeltaStatistic,
+    getRelations,
+    getRelationsDashboard,
+    getCurrent
   };
   const actionMap = {actions: bindActionCreators(actions, dispatch)};
   return actionMap;
-}
+};
 
 export default withRouter(connect(mapStateToProps, mapDispatchToProps)(ProfileScene));
 
@@ -119,11 +131,11 @@ ProfileScene.propTypes = {
     totalCount: PropTypes.number.isRequired
   }).isRequired,
   relationsDashboard: RelationDashboardModel,
-}
+};
 
 const Content = styled.div`
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
 `;