+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Possibility to change email
+- Popup in order to edit personal info
 
+### Removed
+- Variable 'SERVER_HOST'
 
 ## [1.1.0] - 2019-02-22
 ### Added