+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.1] - 2019-03-04
 ### Fixed
 - Round up to 2 decimal places in 'Games lost' field at Profile page.
 
@@ -26,7 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/zensoftio/ZenKicker/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/zensoftio/ZenKicker/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/zensoftio/ZenKicker/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/zensoftio/ZenKicker/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/zensoftio/ZenKicker/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/zensoftio/ZenKicker/releases/tag/v1.0.0